+++ ext/quic/quic.c
@@ -1,24 +1,24 @@
 #include "quic.h"
 #include "stream.h"
 
-VALUE rb_mQuic;
-VALUE rb_mQuicConnection;
-VALUE rb_cQuicConnectionClient;
+VALUE rb_mQUIC;
+VALUE rb_mQUICConnection;
+VALUE rb_cQUICConnectionClient;
 
-VALUE rb_eQuicError;
-VALUE rb_eQuicErrorProto;
-VALUE rb_eQuicErrorDropConn;
-VALUE rb_eQuicErrorRetry;
-VALUE rb_eQuicErrorClosed;
-VALUE rb_eQuicErrorCryptoError;
-VALUE rb_eQuicErrorHandshakeTimeout;
-VALUE rb_eQuicErrorFlowControl;
-VALUE rb_eQuicErrorUnknown;
-VALUE rb_eQuicErrorWaitReadable;
-VALUE rb_eQuicErrorWaitWritable;
-VALUE rb_eQuicErrorStreamClosed;
-VALUE rb_eQuicErrorStreamReset;
-VALUE rb_eQuicErrorNotBound;
+VALUE rb_eQUICError;
+VALUE rb_eQUICErrorProto;
+VALUE rb_eQUICErrorDropConn;
+VALUE rb_eQUICErrorRetry;
+VALUE rb_eQUICErrorClosed;
+VALUE rb_eQUICErrorCryptoError;
+VALUE rb_eQUICErrorHandshakeTimeout;
+VALUE rb_eQUICErrorFlowControl;
+VALUE rb_eQUICErrorUnknown;
+VALUE rb_eQUICErrorWaitReadable;
+VALUE rb_eQUICErrorWaitWritable;
+VALUE rb_eQUICErrorStreamClosed;
+VALUE rb_eQUICErrorStreamReset;
+VALUE rb_eQUICErrorNotBound;
 
 static VALUE
 quic_library_versions(VALUE self)
@@ -42,35 +42,35 @@ quic_raise_ngtcp2_error(int rv)
   VALUE cls;
   switch (rv) {
     case NGTCP2_ERR_PROTO:
-      cls = rb_eQuicErrorProto;
+      cls = rb_eQUICErrorProto;
       break;
     case NGTCP2_ERR_DROP_CONN:
-      cls = rb_eQuicErrorDropConn;
+      cls = rb_eQUICErrorDropConn;
       break;
     case NGTCP2_ERR_RETRY:
-      cls = rb_eQuicErrorRetry;
+      cls = rb_eQUICErrorRetry;
       break;
     case NGTCP2_ERR_CLOSING:
     case NGTCP2_ERR_DRAINING:
-      cls = rb_eQuicErrorClosed;
+      cls = rb_eQUICErrorClosed;
       break;
     case NGTCP2_ERR_IDLE_CLOSE:
-      cls = rb_eQuicErrorClosed;
+      cls = rb_eQUICErrorClosed;
       break;
     case NGTCP2_ERR_CRYPTO:
-      cls = rb_eQuicErrorCryptoError;
+      cls = rb_eQUICErrorCryptoError;
       break;
     case NGTCP2_ERR_HANDSHAKE_TIMEOUT:
-      cls = rb_eQuicErrorHandshakeTimeout;
+      cls = rb_eQUICErrorHandshakeTimeout;
       break;
     case NGTCP2_ERR_STREAM_DATA_BLOCKED:
-      cls = rb_eQuicErrorFlowControl;
+      cls = rb_eQUICErrorFlowControl;
       break;
     case NGTCP2_ERR_STREAM_SHUT_WR:
-      cls = rb_eQuicErrorStreamClosed;
+      cls = rb_eQUICErrorStreamClosed;
       break;
     default:
-      cls = rb_eQuicErrorUnknown;
+      cls = rb_eQUICErrorUnknown;
       break;
   }
 
@@ -82,37 +82,37 @@ quic_raise_ngtcp2_error(int rv)
 RUBY_FUNC_EXPORTED void
 Init_quic(void)
 {
-  rb_mQuic = rb_define_module("Quic");
-  rb_mQuicConnection = rb_define_module_under(rb_mQuic, "Connection");
-  rb_define_singleton_method(rb_mQuic, "library_versions", quic_library_versions, 0);
+  rb_mQUIC = rb_define_module("QUIC");
+  rb_mQUICConnection = rb_define_module_under(rb_mQUIC, "Connection");
+  rb_define_singleton_method(rb_mQUIC, "library_versions", quic_library_versions, 0);
 
-  /* Quic::Error is defined in Ruby (lib/quic.rb) before this Init_quic runs,
+  /* QUIC::Error is defined in Ruby (lib/quic.rb) before this Init_quic runs,
      so rb_const_get retrieves the already-defined base class. Subclasses live
-     under Quic::Error::<Name>. */
-  rb_eQuicError = rb_const_get(rb_mQuic, rb_intern("Error"));
-  rb_eQuicErrorProto = rb_define_class_under(rb_eQuicError, "Proto", rb_eQuicError);
-  rb_eQuicErrorDropConn = rb_define_class_under(rb_eQuicError, "DropConn", rb_eQuicError);
-  rb_eQuicErrorRetry = rb_define_class_under(rb_eQuicError, "Retry", rb_eQuicError);
-  rb_eQuicErrorClosed = rb_define_class_under(rb_eQuicError, "Closed", rb_eQuicError);
-  rb_eQuicErrorCryptoError = rb_define_class_under(rb_eQuicError, "CryptoError", rb_eQuicError);
-  rb_eQuicErrorHandshakeTimeout = rb_define_class_under(rb_eQuicError, "HandshakeTimeout", rb_eQuicError);
-  rb_eQuicErrorFlowControl = rb_define_class_under(rb_eQuicError, "FlowControl", rb_eQuicError);
-  rb_eQuicErrorUnknown = rb_define_class_under(rb_eQuicError, "Unknown", rb_eQuicError);
+     under QUIC::Error::<Name>. */
+  rb_eQUICError = rb_const_get(rb_mQUIC, rb_intern("Error"));
+  rb_eQUICErrorProto = rb_define_class_under(rb_eQUICError, "Proto", rb_eQUICError);
+  rb_eQUICErrorDropConn = rb_define_class_under(rb_eQUICError, "DropConn", rb_eQUICError);
+  rb_eQUICErrorRetry = rb_define_class_under(rb_eQUICError, "Retry", rb_eQUICError);
+  rb_eQUICErrorClosed = rb_define_class_under(rb_eQUICError, "Closed", rb_eQUICError);
+  rb_eQUICErrorCryptoError = rb_define_class_under(rb_eQUICError, "CryptoError", rb_eQUICError);
+  rb_eQUICErrorHandshakeTimeout = rb_define_class_under(rb_eQUICError, "HandshakeTimeout", rb_eQUICError);
+  rb_eQUICErrorFlowControl = rb_define_class_under(rb_eQUICError, "FlowControl", rb_eQUICError);
+  rb_eQUICErrorUnknown = rb_define_class_under(rb_eQUICError, "Unknown", rb_eQUICError);
 
   /* IO-style non-blocking errors. Mix in the standard IO::Wait{Readable,Writable}
-     modules so callers can `rescue IO::WaitReadable` (or the Quic subclass)
+     modules so callers can `rescue IO::WaitReadable` (or the QUIC subclass)
      interchangeably with stdlib IO objects. */
   VALUE io_wait_readable = rb_const_get(rb_cIO, rb_intern("WaitReadable"));
   VALUE io_wait_writable = rb_const_get(rb_cIO, rb_intern("WaitWritable"));
-  rb_eQuicErrorWaitReadable = rb_define_class_under(rb_eQuicError, "WaitReadable", rb_eQuicError);
-  rb_include_module(rb_eQuicErrorWaitReadable, io_wait_readable);
-  rb_eQuicErrorWaitWritable = rb_define_class_under(rb_eQuicError, "WaitWritable", rb_eQuicError);
-  rb_include_module(rb_eQuicErrorWaitWritable, io_wait_writable);
+  rb_eQUICErrorWaitReadable = rb_define_class_under(rb_eQUICError, "WaitReadable", rb_eQUICError);
+  rb_include_module(rb_eQUICErrorWaitReadable, io_wait_readable);
+  rb_eQUICErrorWaitWritable = rb_define_class_under(rb_eQUICError, "WaitWritable", rb_eQUICError);
+  rb_include_module(rb_eQUICErrorWaitWritable, io_wait_writable);
 
-  rb_eQuicErrorStreamClosed = rb_define_class_under(rb_eQuicError, "StreamClosed", rb_eQuicError);
-  rb_eQuicErrorStreamReset = rb_define_class_under(rb_eQuicError, "StreamReset", rb_eQuicError);
-  rb_eQuicErrorNotBound = rb_define_class_under(rb_eQuicError, "NotBound", rb_eQuicError);
+  rb_eQUICErrorStreamClosed = rb_define_class_under(rb_eQUICError, "StreamClosed", rb_eQUICError);
+  rb_eQUICErrorStreamReset = rb_define_class_under(rb_eQUICError, "StreamReset", rb_eQUICError);
+  rb_eQUICErrorNotBound = rb_define_class_under(rb_eQUICError, "NotBound", rb_eQUICError);
 
-  Init_quic_connection_client(rb_mQuicConnection);
-  Init_quic_stream(rb_mQuic);
+  Init_quic_connection_client(rb_mQUICConnection);
+  Init_quic_stream(rb_mQUIC);
 }